requirements.txt
keras==2.13.1
keras==2.12.0
numpy==1.24.3
numpy==1.23.5
protobuf==5.26.0
python-dotenv==1.0.1
streamlit==1.32.2
streamlit_option_menu==0.3.12
